Move utilities into utils folder
